fixing architecture
